pyproject.toml
[tool.poetry]
name = "example2"
version = "0.1.0"
description = "example2"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0.a1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.5.0"
nonebot-plugin-test = "^0.3.0"


[tool.nonebot]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_status", "nonebot_plugin_novelai"]
plugin_dirs = []
[nonebot.plugins]
plugins = ["nonebot_plugin_heweather", "nonebot_plugin_phlogo", "leetcode"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
